docs(router): split policy example stages
