+++ examples/router/custom-policy-example/simple-stacked-score-pick/src/scorer.rs
@@ -1,0 +1,10 @@
+// SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+//! Scorers composed by the `simple-stacked-score-pick` policy.
+
+mod active_requests;
+mod uncached_blocks;
+
+pub(crate) use active_requests::ActiveRequestsScorer;
+pub(crate) use uncached_blocks::UncachedBlocksScorer;
