--- examples/router/custom-policy-example/simple-stacked-score-pick/src/selection.rs
@@ -1,65 +1,0 @@
-// SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-//! Stacked scorers and lowest-cost picker.
-
-use dynamo_kv_router::{
-    WorkerCandidate, WorkerInputView, WorkerInputs, WorkerPicker, WorkerScorer,
-    WorkerSelectionContext, WorkerSelectionPolicyError,
-};
-
-pub(crate) struct ActiveRequestsScorer;
-
-impl WorkerScorer for ActiveRequestsScorer {
-    fn required_worker_inputs(&self) -> WorkerInputs {
-        WorkerInputs::LOAD
-    }
-
-    fn score(
-        &mut self,
-        _context: &WorkerSelectionContext<'_>,
-        candidate: &WorkerCandidate,
-    ) -> Result<f64, WorkerSelectionPolicyError> {
-        let load = candidate
-            .load()
-            .ok_or_else(|| WorkerSelectionPolicyError::failed("load input unavailable"))?;
-        Ok(load.active_requests() as f64)
-    }
-}
-
-pub(crate) struct UncachedBlocksScorer;
-
-impl WorkerScorer for UncachedBlocksScorer {
-    fn required_worker_inputs(&self) -> WorkerInputs {
-        WorkerInputs::CACHE
-    }
-
-    fn score(
-        &mut self,
-        context: &WorkerSelectionContext<'_>,
-        candidate: &WorkerCandidate,
-    ) -> Result<f64, WorkerSelectionPolicyError> {
-        let cache = candidate
-            .cache()
-            .ok_or_else(|| WorkerSelectionPolicyError::failed("cache input unavailable"))?;
-        Ok((context.request_blocks() as f64 - cache.device_overlap_blocks()).max(0.0))
-    }
-}
-
-pub(crate) struct LowestCostPicker;
-
-impl WorkerPicker for LowestCostPicker {
-    fn pick(
-        &mut self,
-        _context: &WorkerSelectionContext<'_>,
-        input: WorkerInputView<'_>,
-    ) -> Result<usize, WorkerSelectionPolicyError> {
-        input
-            .candidates()
-            .iter()
-            .enumerate()
-            .min_by(|(_, left), (_, right)| left.cost().total_cmp(&right.cost()))
-            .map(|(row, _)| row)
-            .ok_or_else(|| WorkerSelectionPolicyError::failed("no eligible worker"))
-    }
-}
